Split Display impls for Type into their own file
